Support intent-only AIs.
In the UI side, it now correctly calls updateTraining to trigger the API to upload the training materials (intents) to the backend.
On the API side it allows training to be updated from AI_UNDEFINED as long as there's no user training file (in that case then the state could never be AI_UNDEFINED unless there was an error uploading the user training file). Updated case where w were returning 500 if the user would call updateTraining on an AI in the wrong state - changed it to Bad Request

Related work items: #2130

+++ service/src/main/java/com/hutoma/api/logic/TrainingLogic.java
@@ -293,22 +293,36 @@ public class TrainingLogic {
                 case AI_READY_TO_TRAIN:     // fallthrough
                 case AI_TRAINING_STOPPED:   // fallthrough
                 case AI_TRAINING_COMPLETE:  // fallthrough
-                case AI_TRAINING_QUEUED:
+                case AI_TRAINING_QUEUED:    // fallthrough
+                case AI_UNDEFINED:
                     try {
                         String trainingMaterials = this.getTrainingMaterialsCommon(devid, aiid);
+                        if (trainingMaterials == null || trainingMaterials.isEmpty()) {
+                            this.logger.logUserTraceEvent(LOGFROM, "UpdateTraining - no training data",
+                                    devidString, logMap);
+                            return ApiError.getNotFound("There is no training data.");
+                        }
+                        // We only support AI_UNDEFINED when it's an intent-only AI (no training file)
+                        if (ai.getSummaryAiStatus() == TrainingStatus.AI_UNDEFINED && ai.trainingFileUploaded()) {
+                            this.logger.logUserTraceEvent(LOGFROM, "UpdateTraining - no training data",
+                                    devidString, logMap);
+                            return ApiError.getBadRequest("Invalid training status - make sure you have uploaded a "
+                                    + "training file and/or added intents.");
+                        }
                         this.aiServices.uploadTraining(ai.getBackendStatus(), devid, aiid, trainingMaterials);
                         // Delete all memory variables for this AI
                         this.memoryIntentHandler.deleteAllIntentsForAi(aiid);
                         this.logger.logUserTraceEvent(LOGFROM, "UpdateTraining", devidString, logMap);
-                        return new ApiResult().setSuccessStatus("Training updated");
+                        return new ApiResult().setSuccessStatus("Training updated.");
                     } catch (AIServices.AiServicesException ex) {
-                        this.logger.logUserExceptionEvent(LOGFROM, "UpdateTraining", devidString, ex);
-                        return ApiError.getInternalServerError("Could not update training");
+                        this.logger.logUserExceptionEvent(LOGFROM, ex.getMessage(), devidString, ex);
+                        return ApiError.getInternalServerError("Could not update training.");
                     }
 
                 default:
-                    this.logger.logUserTraceEvent(LOGFROM, "UpdateTraining - could not update training", devidString, logMap);
-                    return ApiError.getInternalServerError("Could not update the current training session.");
+                    this.logger.logUserTraceEvent(LOGFROM, "UpdateTraining - could not update training",
+                            devidString, logMap);
+                    return ApiError.getBadRequest("Invalid training status");
             }
         } catch (Exception e) {
             this.logger.logUserExceptionEvent(LOGFROM, "UpdateTraining", devidString, e);
